Crate memcached_bpr

Nom based decoder for Memcached Binary Packet Protol

This supports the protocol as of what is in the standard for 1.5 (the 1.6 extensions which are not formally specified are not supported).

Modules

opcode
Binary Packet Protocols

request
RequestHeaders/Packets

response
ResponseHeaders/Packets

status
Status Codes for Response Packets

Enums

ErrorKind
indicates which parser returned an error

IResult
Holds the result of parsing functions

Needed
Contains information on needed data if a parser returned Incomplete

Traits

Encoding
Trait for encoding the value into a packet

Op
Every valid packet has an OpCode

Packet
Trait for all memcached packets

PacketVal
Values encoded within Memcached packets

ResponseStatus
Trait for handling packets which can hold status